requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.5
amqp==5.0.6; python_version >= '3.6'
anyio==3.3.4; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
billiard==3.6.4.0
celery==5.2.1
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
click==8.0.3; python_version >= '3.6'
fastapi==0.70.0
greenlet==1.1.2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.12.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
importlib-metadata==4.8.2; python_version < '3.9'
importlib-resources==5.4.0; python_version < '3.9'
kombu==5.2.2
mako==1.1.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.1; python_version >= '3.6'
prompt-toolkit==3.0.22; python_full_version >= '3.6.2'
psycopg2-binary==2.9.2
pydantic==1.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy==1.4.27
starlette==0.16.0; python_version >= '3.6'
typing-extensions==4.0.0; python_version >= '3.6'
uvicorn==0.15.0
vine==5.0.0; python_version >= '3.6'
wcwidth==0.2.5
zipp==3.6.0; python_version < '3.10'
